% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_nail_catdes.R
\name{shiny_nail_catdes}
\alias{shiny_nail_catdes}
\title{Launch a Shiny app for interpreting a (latent) categorical variable}
\usage{
shiny_nail_catdes(dataset)
}
\arguments{
\item{dataset}{A data frame containing the categorical variable to be analyzed.}
}
\value{
This function does not return a value; it launches a Shiny app.
}
\description{
This function launches a Shiny app for interpreting a (latent)
categorical variable with the 'Nailer' package.
}
\examples{
if(interactive()){
# Processing time is often longer than ten seconds
# because the function uses a large language model.

library(NaileR)
data(iris)
intro_iris <- "A study measured various parts of iris flowers
from 3 different species: setosa, versicolor and virginica.
I will give you the results from this study.
You will have to identify what sets these flowers apart."
intro_iris <- gsub('\n', ' ', intro_iris) |>
stringr::str_squish()
intro_iris

req_iris <- "Please explain what makes each species distinct.
Also, tell me which species has the biggest flowers,
and which species has the smallest."
req_iris <- gsub('\n', ' ', req_iris) |>
stringr::str_squish()
req_iris

shiny_nail_catdes(iris)
}
}
