% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_nail_catdes.R
\name{shiny_nail_catdes}
\alias{shiny_nail_catdes}
\title{Shiny app principale}
\usage{
shiny_nail_catdes(dataset)
}
\arguments{
\item{dataset}{A data frame containing at least one categorical variable (factor).}
}
\value{
Launches a Shiny web application.
}
\description{
Shiny app principale
}
