% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_nail_catdes.R
\name{shiny_nail_catdes}
\alias{shiny_nail_catdes}
\title{Launch a Shiny app for interpreting a (latent) categorical variable}
\usage{
shiny_nail_catdes(dataset)
}
\arguments{
\item{dataset}{A data frame containing the categorical variable to be analyzed.}
}
\description{
Launch a Shiny app for interpreting a (latent) categorical variable
}
