% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_nail_qda.R
\name{shiny_nail_qda}
\alias{shiny_nail_qda}
\title{Launch a Shiny app for analyzing Quantitative Descriptive Analysis data (QDA)}
\usage{
shiny_nail_qda(dataset)
}
\arguments{
\item{dataset}{A data frame containing the data to be analyzed.}
}
\value{
This function does not return a value; it launches a Shiny app.
}
\description{
This function launches a Shiny app for analyzing QDA data with the 'Nailer' package.
The app allows users to explore and analyze their QDA dataset.
}
\examples{
if(interactive()){
# Processing time is often longer than ten seconds
# because the function uses a large language model.

library(SensoMineR)
data(chocolates)
shiny_nail_qda(sensochoc)

}
}
